+++ backend/src/main/java/com/multi/mariage/category/service/DrinkUpperCategoryService.java
@@ -2,9 +2,9 @@ package com.multi.mariage.category.service;
 
 import com.multi.mariage.category.domain.DrinkUpperCategory;
 import com.multi.mariage.category.domain.Region;
-import com.multi.mariage.category.vo.uppercategory.DrinkUpperCategoriesVO;
+import com.multi.mariage.category.vo.drinkupper.DrinkUpperCategoriesVO;
 import com.multi.mariage.category.dto.response.DrinkUpperCategoryResponse;
-import com.multi.mariage.category.vo.uppercategory.DrinkUpperCategoryValuesVO;
+import com.multi.mariage.category.vo.drinkupper.DrinkUpperCategoryValuesVO;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
