+++ backend/src/main/java/com/multi/mariage/category/vo/drinklower/DrinkLowerCategoryValuesVO.java
@@ -1,4 +1,4 @@
-package com.multi.mariage.category.vo.lowercategory;
+package com.multi.mariage.category.vo.drinklower;
 
 import lombok.AccessLevel;
 import lombok.Builder;
