+++ backend/src/test/java/com/multi/mariage/category/service/DrinkLowerCategoryServiceTest.java
@@ -1,8 +1,8 @@
 package com.multi.mariage.category.service;
 
 import com.multi.mariage.category.dto.response.DrinkLowerCategoryResponse;
-import com.multi.mariage.category.vo.lowercategory.DrinkLowerCategoriesVO;
-import com.multi.mariage.category.vo.lowercategory.DrinkLowerCategoryValuesVO;
+import com.multi.mariage.category.vo.drinklower.DrinkLowerCategoriesVO;
+import com.multi.mariage.category.vo.drinklower.DrinkLowerCategoryValuesVO;
 import com.multi.mariage.common.annotation.ServiceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
